Correct formatting of library.properties category field
Previously, the category field was defined on the same line as the paragraph field. This caused the paragraph field to be defined as "category=Device Control" and for the Arduino IDE to display a warning whenever the library was compiled:

WARNING: Category '' in library Arduino-FileUtil is not valid. Setting to 'Uncategorized'